+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - AVIF support.
 - Azure Blob Storage support.
 - `IMGPROXY_STRIP_COLOR_PROFILE` config and [strip_color_profile](https://docs.imgproxy.net/#/generating_the_url_advanced?id=strip-color-profile) processing option.
+- `IMGPROXY_FORMAT_QUALITY` config.
 - (pro) Remove Adobe Illustrator garbage from SVGs.
 
 ### Changed